+++ SoftResListHandling.lua
@@ -587,11 +587,9 @@ function SoftRes.list:getSoftReserves(arg1, arg2)
     local user = string.sub(arg2, 1, string.find(arg2, "-")-1)
     local itemId = nil
     local isIn = false
-    local itemLink = nil
 
     -- We get an itemlink from arg1 and returns the id.
     itemId = SoftRes.helpers:getItemIdFromLink(arg1)
-    itemLink = SoftRes.helpers:getItemLinkFromId(itemId)
 
     -- add the item to the list.
     for i = 1, #SoftResList.players do
@@ -604,16 +602,10 @@ function SoftRes.list:getSoftReserves(arg1, arg2)
         end
     end
 
-    if (not itemLink) then return end
-
     -- If the player is not On the list, we add him
     if (not isIn) then
         SoftRes.list:addSoftReservePlayer(user, itemId)
     end
-    
-    -- Send message ot the player.        
-    local whisperText = "Your SoftReserve of " .. tostring(itemLink) .. " is confirmed. GL HF"
---    ChatThrottleLib:SendChatMessage("NORMAL", "SoftResRollAnnounce", whisperText, "WHISPER", nil, user, nil, nil, nil)
 
     -- update the list.
     SoftRes.list:reOrderPlayerList()